+++ changelog.md
@@ -1,3 +1,5 @@
+- stderr & stdout now get appended to for each stream for remote commands, so you'll get the full response.
+
 # 0.1.2
 - fix occasional LocalJumpError in Remote::SSH#run